Various fixes regarding export options
Renamed ExportOptions.ui and associated header to ui_export_h and created ExportOptions.cpp and its header for showing the dialog box.
