+++ CMakeLists.txt
@@ -31,10 +31,10 @@ include_directories(${SQLITE3_INCLUDE_DIRS})
 message(STATUS ${CMAKE_BUILD_TYPE})
 IF(CMAKE_BUILD_TYPE MATCHES Release)
     message(STATUS "Building release.")
-    add_executable(Ownly WIN32 src/main.cpp src/Database.cpp src/Database.h src/main.h src/exporters.cpp src/exporters.h)
+    add_executable(Ownly WIN32 src/main.cpp src/Database.cpp src/Database.h src/main.h src/exporters.cpp src/exporters.h src/ExportOptions.cpp src/ExportOptions.h)
 ELSE()
     message(STATUS "Building debug.")
-    add_executable(Ownly src/main.cpp src/Database.cpp src/Database.h src/main.h src/exporters.cpp src/exporters.h)
+    add_executable(Ownly src/main.cpp src/Database.cpp src/Database.h src/main.h src/exporters.cpp src/exporters.h src/ExportOptions.cpp src/ExportOptions.h)
 ENDIF()
 
 add_executable(functions_test src/Database.cpp test/test-main.cpp test/test_database_functions.cpp test/test_export_functions.cpp src/exporters.cpp)
