+++ src/ExportOptions.cpp
@@ -1,0 +1,10 @@
+//
+// Created by rwardrup on 9/4/2020.
+//
+
+#include "ExportOptions.h"
+#include "ui_export_options.h"
+
+ExportDialog::ExportDialog(QWidget *parent) {
+    setupUi(this);
+}
